+++ tests/archives/archive_base_test.py
@@ -46,15 +46,20 @@ def test_iteration():
         assert elite.metadata == data.metadata
 
 
-def test_add_during_iteration():
+def test_add_during_iteration(add_mode):
     # Even with just one entry, adding during iteration should still raise an
     # error, just like it does in set.
     data = get_archive_data("GridArchive")
     with pytest.raises(RuntimeError):
         for _ in data.archive_with_elite:
-            data.archive_with_elite.add_single(data.solution,
-                                               data.objective + 1,
-                                               data.measures)
+            if add_mode == "single":
+                data.archive_with_elite.add_single(data.solution,
+                                                   data.objective + 1,
+                                                   data.measures)
+            else:
+                data.archive_with_elite.add([data.solution],
+                                            [data.objective + 1],
+                                            [data.measures])
 
 
 def test_clear_during_iteration():
@@ -64,14 +69,19 @@ def test_clear_during_iteration():
             data.archive_with_elite.clear()
 
 
-def test_clear_and_add_during_iteration():
+def test_clear_and_add_during_iteration(add_mode):
     data = get_archive_data("GridArchive")
     with pytest.raises(RuntimeError):
         for _ in data.archive_with_elite:
             data.archive_with_elite.clear()
-            data.archive_with_elite.add_single(data.solution,
-                                               data.objective + 1,
-                                               data.measures)
+            if add_mode == "single":
+                data.archive_with_elite.add_single(data.solution,
+                                                   data.objective + 1,
+                                                   data.measures)
+            else:
+                data.archive_with_elite.add([data.solution],
+                                            [data.objective + 1],
+                                            [data.measures])
 
 
 #
@@ -90,13 +100,19 @@ def test_stats_dtype(dtype):
     assert isinstance(data.archive_with_elite.stats.obj_mean, dtype)
 
 
-def test_stats_multiple_add():
+def test_stats_multiple_add(add_mode):
     archive = GridArchive(solution_dim=3,
                           dims=[10, 20],
                           ranges=[(-1, 1), (-2, 2)])
-    archive.add_single([1, 2, 3], 1.0, [0, 0])
-    archive.add_single([1, 2, 3], 2.0, [0.25, 0.25])
-    archive.add_single([1, 2, 3], 3.0, [-0.25, -0.25])
+    if add_mode == "single":
+        archive.add_single([1, 2, 3], 1.0, [0, 0])
+        archive.add_single([1, 2, 3], 2.0, [0.25, 0.25])
+        archive.add_single([1, 2, 3], 3.0, [-0.25, -0.25])
+    else:
+        solution_batch = [[1, 2, 3]] * 3
+        objective_batch = [1.0, 2.0, 3.0]
+        measures_batch = [[0, 0], [0.25, 0.25], [-0.25, -0.25]]
+        archive.add(solution_batch, objective_batch, measures_batch)
 
     assert archive.stats.num_elites == 3
     assert np.isclose(archive.stats.coverage, 3 / 200)
@@ -105,15 +121,21 @@ def test_stats_multiple_add():
     assert np.isclose(archive.stats.obj_mean, 2.0)
 
 
-def test_stats_add_and_overwrite():
+def test_stats_add_and_overwrite(add_mode):
     archive = GridArchive(solution_dim=3,
                           dims=[10, 20],
                           ranges=[(-1, 1), (-2, 2)])
-    archive.add_single([1, 2, 3], 1.0, [0, 0])
-    archive.add_single([1, 2, 3], 2.0, [0.25, 0.25])
-    archive.add_single([1, 2, 3], 3.0, [-0.25, -0.25])
-    archive.add_single([1, 2, 3], 5.0,
-                       [0.25, 0.25])  # Overwrites the second add.
+    if add_mode == "single":
+        archive.add_single([1, 2, 3], 1.0, [0, 0])
+        archive.add_single([1, 2, 3], 2.0, [0.25, 0.25])
+        archive.add_single([1, 2, 3], 3.0, [-0.25, -0.25])
+        archive.add_single([1, 2, 3], 5.0,
+                           [0.25, 0.25])  # Overwrites the second add.
+    else:
+        solution_batch = [[1, 2, 3]] * 4
+        objective_batch = [1.0, 2.0, 3.0, 5.0]
+        measures_batch = [[0, 0], [0.25, 0.25], [-0.25, -0.25], [0.25, 0.25]]
+        archive.add(solution_batch, objective_batch, measures_batch)
 
     assert archive.stats.num_elites == 3
     assert np.isclose(archive.stats.coverage, 3 / 200)
@@ -122,7 +144,7 @@ def test_stats_add_and_overwrite():
     assert np.isclose(archive.stats.obj_mean, 3.0)
 
 
-def test_best_elite():
+def test_best_elite(add_mode):
     archive = GridArchive(solution_dim=3,
                           dims=[10, 20],
                           ranges=[(-1, 1), (-2, 2)])
@@ -131,7 +153,10 @@ def test_best_elite():
     assert archive.best_elite is None
 
     # Add an elite.
-    archive.add_single([1, 2, 3], 1.0, [0, 0])
+    if add_mode == "single":
+        archive.add_single([1, 2, 3], 1.0, [0, 0])
+    else:
+        archive.add([[1, 2, 3]], [1.0], [[0, 0]])
 
     assert np.isclose(archive.best_elite.solution, [1, 2, 3]).all()
     assert np.isclose(archive.best_elite.objective, 1.0).all()
@@ -139,7 +164,10 @@ def test_best_elite():
 
     # Add an elite into the same cell as the previous elite -- best_elite should
     # now be overwritten.
-    archive.add_single([4, 5, 6], 2.0, [0, 0])
+    if add_mode == "single":
+        archive.add_single([4, 5, 6], 2.0, [0, 0])
+    else:
+        archive.add([[4, 5, 6]], [2.0], [[0, 0]])
 
     assert np.isclose(archive.best_elite.solution, [4, 5, 6]).all()
     assert np.isclose(archive.best_elite.objective, 2.0).all()
@@ -206,6 +234,14 @@ def test_measure_dim_correct(data):
 
 def test_solution_dim_correct(data):
     assert data.archive.solution_dim == len(data.solution)
+
+
+def test_learning_rate_correct(data):
+    assert data.archive.learning_rate == 1.0  # Default value.
+
+
+def test_threshold_min_correct(data):
+    assert data.archive.threshold_min == -np.inf  # Default value.
 
 
 def test_basic_stats(data):