+++ tests/emitters/gradient_aborescence_emitter_test.py
@@ -7,15 +7,25 @@ from ribs.emitters import GradientAborescenceEmitter
 
 
 def test_auto_batch_size():
-    archive = GridArchive(10, [20, 20], [(-1.0, 1.0)] * 2)
-    emitter = GradientAborescenceEmitter(archive, np.zeros(10), 1.0, 1.0)
+    archive = GridArchive(solution_dim=10,
+                          dims=[20, 20],
+                          ranges=[(-1.0, 1.0)] * 2)
+    emitter = GradientAborescenceEmitter(archive,
+                                         x0=np.zeros(10),
+                                         sigma0=1.0,
+                                         step_size=1.0)
     assert emitter.batch_size is not None
     assert isinstance(emitter.batch_size, int)
 
 
 def test_list_as_initial_solution():
-    archive = GridArchive(10, [20, 20], [(-1.0, 1.0)] * 2)
-    emitter = GradientAborescenceEmitter(archive, [0.0] * 10, 1.0, 1.0)
+    archive = GridArchive(solution_dim=10,
+                          dims=[20, 20],
+                          ranges=[(-1.0, 1.0)] * 2)
+    emitter = GradientAborescenceEmitter(archive,
+                                         x0=[0.0] * 10,
+                                         sigma0=1.0,
+                                         step_size=1.0)
 
     # The list was passed in but should be converted to a numpy array.
     assert isinstance(emitter.x0, np.ndarray)
@@ -25,6 +35,12 @@ def test_list_as_initial_solution():
 @pytest.mark.parametrize("dtype", [np.float64, np.float32],
                          ids=["float64", "float32"])
 def test_dtypes(dtype):
-    archive = GridArchive(10, [20, 20], [(-1.0, 1.0)] * 2, dtype=dtype)
-    emitter = GradientAborescenceEmitter(archive, np.zeros(10), 1.0, 1.0)
+    archive = GridArchive(solution_dim=10,
+                          dims=[20, 20],
+                          ranges=[(-1.0, 1.0)] * 2,
+                          dtype=dtype)
+    emitter = GradientAborescenceEmitter(archive,
+                                         x0=np.zeros(10),
+                                         sigma0=1.0,
+                                         step_size=1.0)
     assert emitter.x0.dtype == dtype